pyproject.toml
[project]
name = "bill-splitter-bot"
version = "0.1.0"
description = "A Telegram bot that intelligently splits restaurant bills using AI vision"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==21.10",
    "google-generativeai==0.8.3",
    "Pillow==11.3.0",
    "qrcode==7.4.2",
    "python-dotenv==1.0.0",
    "aiohttp==3.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
